# vafund-backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk aplikasi VaFund menggunakan Go Fiber dan Hyperledger Fabric",
        "title": "VaFund Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@vafund.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/donations": {
            "get": {
                "description": "Get a list of all donations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Get all donations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DonationResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new donation with sender name, amount and optional message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Create a new donation",
                "parameters": [
                    {
                        "description": "Donation data",
                        "name": "donation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateDonationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.DonationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/donations/by-amount": {
            "get": {
                "description": "Get a list of donations that have at least the specified amount",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Get donations by minimum amount",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Minimum donation amount",
                        "name": "minAmount",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DonationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/donations/event/{eventCode}": {
            "get": {
                "description": "Get a list of donations associated with a specific event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Get donations by event code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Code",
                        "name": "eventCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DonationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/donations/event/{eventCode}/total": {
            "get": {
                "description": "Get total amount and count of donations for a specific event with current amount after withdrawals",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Get total donation statistics by event code with current amount (net of withdrawals)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Code",
                        "name": "eventCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TotalDonationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/donations/total": {
            "get": {
                "description": "Get total amount and count of all donations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Get total donation statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TotalDonationResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/donations/{id}": {
            "get": {
                "description": "Get a specific donation by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Get donation by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Donation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DonationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateDonationRequest": {
            "type": "object",
            "required": [
                "amount",
                "donationId",
                "senderName"
            ],
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "100000"
                },
                "donationId": {
                    "type": "string",
                    "example": "donation123"
                },
                "eventCode": {
                    "type": "string",
                    "example": "RAMADAN2025"
                },
                "message": {
                    "type": "string",
                    "example": "Semoga bermanfaat"
                },
                "senderName": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
...
        "models.DonationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Donation created successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Detailed error description"
                },
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.TotalDonationResponse": {
            "type": "object",
            "properties": {
                "currentAmount": {
                    "type": "number",
                    "example": 300000
                },
                "message": {
                    "type": "string",
                    "example": "Total donations retrieved successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "totalAmount": {
                    "type": "number",
                    "example": 350000
                },
                "totalCount": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    }
}